Ещё enum-ы

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->string('email', 64)->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password', 512);
-            $table->integer('role', unsigned: true)->default(Role::User());
+            $table->mediumInteger('role', unsigned: true)->default(Role::User());
             $table->rememberToken();
             $table->timestamps();
         });
